Commit current version has many things correct
Committed  to save

+++ Iterator.java
@@ -1,4 +1,4 @@
 public interface Iterator<T> extends Iterable<T> {
     public Iterator<T> getNeighborIterator();
-    public Iterator<Integer> getWeightIterator();
+    public Iterator<Double> getWeightIterator();
 }//end Iterator
